Changed Torrent types from u32 to i64 to fix serde issue

+++ src/data.rs
@@ -69,12 +69,12 @@ pub struct MainData {
 
 #[derive(Debug, Deserialize, Getters, Clone)]
 pub struct Torrent {
-    added_on: u32,
+    added_on: i64,
     amount_left: u64,
     auto_tmm: bool,
     category: String,
     completed: i64,
-    completion_on: u32,
+    completion_on: i64,
     dl_limit: i64,
     dlspeed: i64,
     downloaded: i64,
